+++ DitherCalculator/DitherCalculatorApp.swift
@@ -27,8 +27,8 @@ public struct FocalLength: Equatable {
     public static let unit: Self.Unit = .millimeters
     public let measurement: Measurement<Self.Unit>
 
-    public init(value: Double) {
-        measurement = Measurement(value: value, unit: Self.unit)
+    public init(value: Int) {
+        measurement = Measurement(value: Double(value), unit: Self.unit)
     }
 }
 
@@ -46,7 +46,7 @@ struct EquipmentMetadata: Equatable {
     let focalLength: FocalLength
     let pixelSize: PixelSize
 
-    init(focalLength: Double, pixelSize: Double) {
+    init(focalLength: Int, pixelSize: Double) {
         self.focalLength = FocalLength(value: focalLength)
         self.pixelSize = PixelSize(value: pixelSize)
     }
